[f11][171203-2226] modif drive.py because OS installed in Europe (, vs .)
I've modified steering_angle replacing '.' by ','. At the first look, the car behaved better but in fact the angle was switching from 25deg to -25deg and none values in between.

I've also modified in the 2nd time the speed, and the throttle. But this change generated bugs: 
'
   File "drive.py", line 85, in telemetry
    throttle = controller.update(float(speed))
ValueError: could not convert string to float: '0,0000' 
' 

I decided to let it like that and run the training like this. I don't see why the simulator could not handle the fact that the program is run in Europe. Python and other softwares seem not to be affected.